# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI-powered couples therapy backend API",
        "title": "Mend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "mend-backend-j0qd.onrender.com",
    "basePath": "/api",
    "paths": {
        "/api/insights/{userId}": {
            "get": {
                "description": "Returns sessions and reflections submitted by or about the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Insights"
                ],
                "summary": "Get communication insights for a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/invite": {
            "post": {
                "description": "Links two users as partners by ID and stores inviter.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Link partners",
                "parameters": [
                    {
                        "description": "Invite Info (yourId and partnerId)",
                        "name": "invite",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "Logs in user by email lookup.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Login a user",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/moderate": {
            "post": {
                "description": "Returns AI feedback + tone warning",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "GPT-4 AI moderation of transcript",
                "parameters": [
                    {
                        "description": "Transcript, Speaker",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/post-resolution": {
            "post": {
                "description": "Stores emotional bonding data after a session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reflection"
                ],
                "summary": "Save post-resolution reflection",
                "parameters": [
                    {
                        "description": "Post-resolution data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PostResolution"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/reflection": {
            "post": {
                "description": "Stores a post-session reflection entry from a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reflections"
                ],
                "summary": "Submit a session reflection",
                "parameters": [
                    {
                        "description": "Reflection Data",
                        "name": "reflection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Reflection"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Reflection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "description": "Creates a new user account in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User Info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/score": {
            "post": {
                "description": "Partner submits feedback for a session (used for insights)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Communication"
                ],
                "summary": "Submit communication score",
                "parameters": [
                    {
                        "description": "Score Data",
                        "name": "score",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CommunicationScore"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/session": {
            "post": {
                "description": "Creates a session document in MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session"
                ],
                "summary": "Start a new session between partners",
                "parameters": [
                    {
                        "description": "Session Info",
                        "name": "session",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Session"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Session"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "description": "Returns user information by user ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CommunicationScore": {
            "type": "object",
            "properties": {
                "clarity": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "integer"
                },
                "empathy": {
                    "type": "integer"
                },
                "listening": {
                    "type": "integer"
                },
                "openMindedness": {
                    "type": "integer"
                },
                "partnerId": {
                    "description": "Who submitted this",
                    "type": "string"
                },
                "respect": {
                    "type": "integer"
                },
                "responsiveness": {
                    "type": "integer"
                },
                "sessionId": {
                    "description": "Which session this score is for",
                    "type": "string"
                }
            }
        },
        "models.Message": {
            "type": "object",
            "properties": {
                "speakerId": {
                    "description": "Who spoke",
                    "type": "string"
                },
                "text": {
                    "description": "Transcript",
                    "type": "string"
                },
                "timestamp": {
                    "description": "Unix time",
                    "type": "integer"
                }
            }
        },
        "models.PostResolution": {
            "type": "object",
            "properties": {
                "bondingActivity": {
                    "description": "e.g., \"Go for a walk together\"",
                    "type": "string"
                },
                "gratitude": {
                    "description": "e.g., \"I'm grateful for your honesty\"",
                    "type": "string"
                },
                "reflection": {
                    "description": "e.g., \"I felt better after the session\"",
                    "type": "string"
                },
                "sessionId": {
                    "description": "Associated session ID",
                    "type": "string"
                },
                "timestamp": {
                    "description": "Unix time of submission",
                    "type": "integer"
                },
                "userId": {
                    "description": "ID of the user submitting",
                    "type": "string"
                }
            }
        },
        "models.Reflection": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Composite ID: sessionID-userID",
                    "type": "string"
                },
                "sessionId": {
                    "description": "Which session",
                    "type": "string"
                },
                "text": {
                    "description": "Reflection content",
                    "type": "string"
                },
                "timestamp": {
                    "description": "Unix time",
                    "type": "integer"
                },
                "userId": {
                    "description": "Who submitted",
                    "type": "string"
                }
            }
        },
        "models.Session": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "Session time",
                    "type": "integer"
                },
                "id": {
                    "description": "UUID",
                    "type": "string"
                },
                "messages": {
                    "description": "Chat transcript",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Message"
                    }
                },
                "partnerA": {
                    "description": "User A",
                    "type": "string"
                },
                "partnerB": {
                    "description": "User B",
                    "type": "string"
                },
                "resolved": {
                    "description": "Has reflection happened",
                    "type": "boolean"
                },
                "scoreA": {
                    "description": "A's score",
                    "$ref": "#/definitions/models.CommunicationScore"
                },
                "scoreB": {
                    "description": "B's score",
                    "$ref": "#/definitions/models.CommunicationScore"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "challenges": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "colorCode": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "goals": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "invitedBy": {
                    "description": "✅ NEW",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "partnerId": {
                    "type": "string"
                }
            }
        }
    }
}